requirements.txt
APScheduler==3.5.1
asn1crypto==0.24.0
astroid==1.6.2
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cycler==0.10.0
decorator==4.2.1
Flask==0.12.2
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.7.1
httplib2==0.11.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
networkx==2.1
newsapi-python==0.0.2
numpy==1.14.2
oauth2client==4.1.2
pbr==4.0.0
pipenv==11.9.0
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pydub==0.21.0
pylint==1.8.3
pymessenger==0.0.7.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
requests==2.13.0
requests-toolbelt==0.8.0
rsa==3.4.2
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.7
stevedore==1.28.0
Twisted==17.9.0
txaio==2.9.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
watson-developer-cloud==1.2.1
Werkzeug==0.14.1
wrapt==1.10.11
zope.interface==4.4.3
